Type feature: increase task limit to 500.
Although the Fargete Task limit has been increased from 100 to 1000, task_runner can only run 10 tasks in a loop, so each loop needs to wait for 10 seconds in order to avoid throttle. Therefore, we also need to increase the execution time of task-runner to the maximum limit of 900 seconds. In order to avoid Lambda execution timeout, I suggest raising the upper limit to 500 Tasks is a safe value.